priorities/serializers.py
from rest_framework import serializers

from priorities.models import Priority


class PrioritySerializer(serializers.ModelSerializer):
    class Meta:
        model = Priority
        fields = "__all__"
