requirement.txt
asgiref==3.8.1
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
crashtest==0.4.1
cryptography==43.0.3
distlib==0.3.9
Django==5.1.2
django-stubs==5.1.0
django-stubs-ext==5.1.0
djangorestframework==3.15.2
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.16.1
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
keyring==24.3.1
more-itertools==10.5.0
msgpack==1.1.0
mypy==1.11.2
mypy-extensions==1.0.0
packaging==24.1
pexpect==4.9.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycparser==2.22
pyproject_hooks==1.2.0
python-dotenv==1.0.1
RapidFuzz==3.10.1
requests==2.32.3
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
sqlparse==0.5.1
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.0
zipp==3.20.2
